Move implementation of schemaTypeQuickPick().
